+++ R/Rules-methods.R
@@ -1093,7 +1093,7 @@ setMethod(
     # Determine what is the basis level for increment,
     # i.e. the last dose or the max dose applied.
     basis_dose_level <- ifelse(
-      increments@basisLevel == "last",
+      increments@basis_level == "last",
       tail(
         data@xLevel,
         1
@@ -1103,7 +1103,7 @@ setMethod(
 
     max_next_dose_level <- min(
       length(data@doseGrid),
-      basis_dose_level + increments@maxLevels
+      basis_dose_level + increments@max_levels
     )
 
     data@doseGrid[max_next_dose_level]